move QTextEditLogHandler to gt_gui module
